+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - 
 
+### v1.37.0
+
+- Add support of wallet export as a JSON backup file
+
 ### v1.36.0
 
 -  Update GTM ID from GTM-MJG54ZJ to GTM-WQ33L7GG